refactor: profile picture
