+++ instapot/src/profile/profileedit.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import EditProfilePicture from "./comp/EditProfilePicture";
 
 const ProfileEdit = () => {
   const posts = {
@@ -21,7 +21,7 @@ const ProfileEdit = () => {
           <div className='flex items-center w-4/6'>
             <div className='flex flex-col gap-1'>
               <span>{posts.username}</span>
-              <EditProfileImage profileImage={posts.profileImage} />
+              <EditProfilePicture profileImage={posts.profileImage} />
             </div>
           </div>
         </div>
@@ -55,35 +55,6 @@ const ProfileEdit = () => {
       </div>
     </div>
   );
-};
-
-type TEditProfileImage = {
-  profileImage: string;
-};
-
-const EditProfileImage = ({ profileImage }: TEditProfileImage) => {
-  const [changeProfileImage, setChangeProfileImage] = useState<Boolean>(false);
-
-  const hidden = (
-    <>
-      <button className='font-bold' onClick={() => setChangeProfileImage(true)}>
-        Ändra profilbild
-      </button>
-    </>
-  );
-
-  const show = (
-    <>
-      <div className='flex flex-row gap-2'>
-        <input type='text' placeholder={profileImage} className='bg-gray-200 rounded-sm px-2 py-1 w-5/6' disabled />
-        <button className='px-4 py-1 rounded-md bg-green-700/70 hover:bg-green-600/70 text-white' onClick={() => setChangeProfileImage(false)}>
-          Save
-        </button>
-      </div>
-    </>
-  );
-
-  return changeProfileImage ? show : hidden;
 };
 
 export default ProfileEdit;
